+++ packages/popover/src/use-popover-init.ts
@@ -25,12 +25,10 @@
  */
 import { ref } from 'vue';
 
-import _ from 'lodash';
-
 import { EMIT_EVENTS } from './const';
 import useFloating from './use-floating';
 import usePopperId from './use-popper-id';
-import { getFullscreenUid, SharedState, random } from './utils';
+import { getFullscreenUid, ReferenceClickSharedState, random } from './utils';
 
 export default (props, ctx, { refReference, refContent, refArrow, refRoot }) => {
   let storeEvents = null;
@@ -206,21 +204,31 @@ export default (props, ctx, { refReference, refContent, refArrow, refRoot }) => 
     document.body.removeEventListener('fullscreenchange', handleFullscreenChange);
   };
 
-  const handleClickOutside = _.debounce((_e: MouseEvent) => {
-    if (SharedState[uniqKey]) {
-      SharedState[uniqKey] = false;
-      return;
-    }
-    ctx.emit(EMIT_EVENTS.CLICK_OUTSIDE, { isShow: localIsShow.value, event: _e });
-    const needExec = props.disableOutsideClick || props.always || props.disabled || props.trigger === 'manual';
-    if (!props.forceClickoutside && needExec) {
-      return;
-    }
+  const handleClickOutside = (_e: MouseEvent, hideIgnoreReference = false) => {
+    const commonFunc = () => {
+      ctx.emit(EMIT_EVENTS.CLICK_OUTSIDE, { isShow: localIsShow.value, event: _e });
+      const needExec = props.disableOutsideClick || props.always || props.disabled || props.trigger === 'manual';
+      if (!props.forceClickoutside && needExec) {
+        return;
+      }
 
-    if (localIsShow.value) {
-      hideFn();
+      if (localIsShow.value) {
+        hideFn();
+      }
+    };
+    if (hideIgnoreReference) {
+      setTimeout(() => {
+        if (ReferenceClickSharedState[uniqKey]) {
+          ReferenceClickSharedState[uniqKey] = false;
+          return;
+        }
+
+        return commonFunc();
+      });
+    } else {
+      return commonFunc();
     }
-  }, 10);
+  };
 
   return {
     onMountedFn,