+++ packages/tree/src/use-node-attribute.tsx
@@ -24,7 +24,7 @@
  * IN THE SOFTWARE.
  */
 
-import { isElement } from 'lodash';
+import isElement from 'lodash/isElement';
 
 import { NODE_ATTRIBUTES, NODE_SOURCE_ATTRS } from './constant';
 import { TreeNode, TreePropTypes } from './props';
@@ -177,11 +177,15 @@ export default (
   const getSourceNodeByUID = (uid: string) => flatData.data.find(item => getNodeId(item) === uid);
 
   const getParentNodeData = (node: TreeNode | string) => {
-    if (isRootNode(node)) {
+    let target = node;
+    if (typeof target === 'string') {
+      target = getSourceNodeByUID(target);
+    }
+    if (isRootNode(target)) {
       return { [props.children]: props.data };
     }
 
-    return getParentNode(node as TreeNode);
+    return getParentNode(target);
   };
 
   /**