+++ site/views/pop-confirm/hide-ignore-reference.vue
@@ -3,6 +3,9 @@
     <bk-pop-confirm
       ref="popConfirmRef"
       width="288"
+      :popover-options="{
+        hideIgnoreReference: true,
+      }"
       content="删除操作无法撤回，请谨慎操作！"
       title="确认删除该脚本？"
       trigger="click"
@@ -10,7 +13,7 @@
       @after-show="handleAfterShow"
       @confirm="ensure"
     >
-      <bk-button>删除</bk-button>
+      <bk-button>删除(点我不收起弹窗)</bk-button>
     </bk-pop-confirm>
   </div>
 </template>